+++ artifacts/api-server/src/models/sidebarPermissionModel.js
@@ -2,11 +2,11 @@ const mongoose = require('mongoose');
 
 const sidebarPermissionSchema = new mongoose.Schema(
   {
-    roleId: { type: mongoose.Schema.Types.ObjectId, ref: 'Role', required: true },
-    industryId: { type: mongoose.Schema.Types.ObjectId, ref: 'Industry', required: true },
-    menuId: { type: mongoose.Schema.Types.ObjectId, ref: 'SidebarMenu', required: true },
-    isVisible: { type: Boolean, default: true },
-    orderOverride: { type: Number, default: null },
+    role_id: { type: mongoose.Schema.Types.ObjectId, ref: 'Role', required: true, alias: 'roleId' },
+    industry_id: { type: mongoose.Schema.Types.ObjectId, ref: 'Industry', required: true, alias: 'industryId' },
+    menu_id: { type: mongoose.Schema.Types.ObjectId, ref: 'SidebarMenu', required: true, alias: 'menuId' },
+    is_visible: { type: Boolean, default: true, alias: 'isVisible' },
+    order_override: { type: Number, default: null, alias: 'orderOverride' },
   },
   { 
     timestamps: true,
@@ -15,24 +15,12 @@ const sidebarPermissionSchema = new mongoose.Schema(
   },
 );
 
-sidebarPermissionSchema.virtual('menu_id')
-  .get(function() { return this.menuId; })
-  .set(function(v) { this.menuId = v; });
-
-sidebarPermissionSchema.virtual('is_visible')
-  .get(function() { return this.isVisible; })
-  .set(function(v) { this.isVisible = v; });
-
-sidebarPermissionSchema.virtual('order_override')
-  .get(function() { return this.orderOverride; })
-  .set(function(v) { this.orderOverride = v; });
-
 sidebarPermissionSchema.index(
-  { roleId: 1, industryId: 1, menuId: 1 },
+  { role_id: 1, industry_id: 1, menu_id: 1 },
   { unique: true, name: 'idx_perm_unique' },
 );
 sidebarPermissionSchema.index(
-  { roleId: 1, industryId: 1, isVisible: 1 },
+  { role_id: 1, industry_id: 1, is_visible: 1 },
   { name: 'idx_perm_lookup' },
 );
 
@@ -47,70 +35,70 @@ exports.SidebarPermission = SidebarPermission;
 exports.list = async ({ roleId, industryId, menuId, menu_id, visibleOnly = false } = {}) => {
   const mId = menuId || menu_id;
   const q = {};
-  if (roleId) q.roleId = roleId;
-  if (industryId) q.industryId = industryId;
-  if (mId) q.menuId = mId;
-  if (visibleOnly) q.isVisible = true;
-  return SidebarPermission.find(q).lean().exec();
+  if (roleId) q.role_id = roleId;
+  if (industryId) q.industry_id = industryId;
+  if (mId) q.menu_id = mId;
+  if (visibleOnly) q.is_visible = true;
+  return SidebarPermission.find(q).exec();
 };
 
-exports.findById = async (id) => SidebarPermission.findById(id).lean().exec();
+exports.findById = async (id) => SidebarPermission.findById(id).exec();
 
 exports.upsert = async ({ roleId, industryId, menuId, menu_id, isVisible, is_visible, orderOverride, order_override }) => {
   const mId = menuId || menu_id;
   const vis = isVisible !== undefined ? isVisible : is_visible;
   const ord = orderOverride !== undefined ? orderOverride : order_override;
   const $set = {};
-  if (vis !== undefined) $set.isVisible = !!vis;
+  if (vis !== undefined) $set.is_visible = !!vis;
   if (ord !== undefined) {
-    $set.orderOverride = ord === null ? null : Number(ord);
+    $set.order_override = ord === null ? null : Number(ord);
   }
   await SidebarPermission.updateOne(
-    { roleId, industryId, menuId: mId },
-    { $set, $setOnInsert: { roleId, industryId, menuId: mId } },
+    { role_id: roleId, industry_id: industryId, menu_id: mId },
+    { $set, $setOnInsert: { role_id: roleId, industry_id: industryId, menu_id: mId } },
     { upsert: true },
   );
-  return SidebarPermission.findOne({ roleId, industryId, menuId: mId }).lean().exec();
+  return SidebarPermission.findOne({ role_id: roleId, industry_id: industryId, menu_id: mId }).exec();
 };
 
-exports.remove = async (id) => SidebarPermission.findByIdAndDelete(id).lean().exec();
+exports.remove = async (id) => SidebarPermission.findByIdAndDelete(id).exec();
 
 exports.removeByCombo = async ({ roleId, industryId, menuId, menu_id }) => {
   const mId = menuId || menu_id;
-  return SidebarPermission.deleteOne({ roleId, industryId, menuId: mId }).exec();
+  return SidebarPermission.deleteOne({ role_id: roleId, industry_id: industryId, menu_id: mId }).exec();
 };
 
 exports.removeByRoleIndustry = async ({ roleId, industryId }) =>
-  SidebarPermission.deleteMany({ roleId, industryId }).exec();
+  SidebarPermission.deleteMany({ role_id: roleId, industry_id: industryId }).exec();
 
 exports.removeByIndustry = async (industryId) =>
-  SidebarPermission.deleteMany({ industryId }).exec();
+  SidebarPermission.deleteMany({ industry_id: industryId }).exec();
 
 exports.removeByRole = async (roleId) =>
-  SidebarPermission.deleteMany({ roleId }).exec();
+  SidebarPermission.deleteMany({ role_id: roleId }).exec();
 
 exports.removeByMenu = async (menuId) =>
-  SidebarPermission.deleteMany({ menuId }).exec();
+  SidebarPermission.deleteMany({ menu_id: menuId }).exec();
 
 exports.bulkSetForRoleIndustry = async ({ roleId, industryId, menuIds, menu_ids }) => {
   const ids = Array.isArray(menuIds || menu_ids) ? (menuIds || menu_ids) : [];
   await SidebarPermission.deleteMany({
-    roleId,
-    industryId,
-    menuId: { $nin: ids },
+    role_id: roleId,
+    industry_id: industryId,
+    menu_id: { $nin: ids },
   });
   if (ids.length) {
     const ops = ids.map((menuId) => ({
       updateOne: {
-        filter: { roleId, industryId, menuId },
+        filter: { role_id: roleId, industry_id: industryId, menu_id: menuId },
         update: {
-          $set: { isVisible: true },
-          $setOnInsert: { roleId, industryId, menuId },
+          $set: { is_visible: true },
+          $setOnInsert: { role_id: roleId, industry_id: industryId, menu_id: menuId },
         },
         upsert: true,
       },
     }));
     await SidebarPermission.bulkWrite(ops, { ordered: false });
   }
-  return SidebarPermission.find({ roleId, industryId }).lean().exec();
+  return SidebarPermission.find({ role_id: roleId, industry_id: industryId }).exec();
 };
